Added the logic behind the slideshow
Changes made in the backend and frontend to ensure that the slideshow
runs as intended. Changes include: updated code for the popular_products
endpoint, added the logic in the homepage_code.js file and slightly
altered the html, css and swagger files.

Clean up is needed as well as changing the structure of the homepage